Update query parsing code to be more manual, but slightly more robust; fixes #22, fixes #24, and fixes #27
Previously, the query parser let `strchr` do the heavy lifting for finding separator tokens. However, this limited the parser to finding only one token when it needed to be able to find any (perhaps for keys without values or for alternating field separators '&' and ';'). Now, we'll just do the iteration ourselves in constant time. This change also uncovered an issue in the function `luadb_strndup` which prohibited that function from performing duplication of strings that were part of larger buffers. Since it used `strncpy`, that function failed to NUL terminate strings which came from a larger buffer. Now, `luadb_strndup` uses `memcpy` and manually adds the NUL terminator.