Improve logging for LuaDB running as a FastCGI process; fixes #9
Use syslog on non-Windows systems to log FastCGI errors and warnings to help administrators diagnose issues in their setup and to help me diagnose bugs. Provide a faux-syslog interface for Windows which merely uses `fprintf` on a log file for now. Move the `PUSH_VAL_STRING` macro into it's own inline function to improve type safety and... just be cleaner.